fileserver/requirements.txt
aioboto3==12.4.0
aiobotocore==2.12.3
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
boto3==1.34.69
botocore==1.34.69
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
yarl==1.9.4
